fundamentals-of-data-visualization/final/requirements.txt
altair==5.0.1
altair-data-server==0.4.1
attrs==23.1.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
parse==1.19.1
portpicker==1.5.2
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
toolz==0.12.0
tornado==6.3.2
tzdata==2023.3
